# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Efrica website  OpenApi  Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "khumbozimba4@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1"
        }
    ],
    "paths": {
...
        "/service/getAllServices": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "Get list of services",
                "description": "Returns list of services",
                "operationId": "getAllServices",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ServiceResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/service/addService": {
            "post": {
                "tags": [
                    "Services"
                ],
                "summary": "Store new service",
                "description": "Returns service data",
                "operationId": "storeService",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreServiceRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Service"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/service/{id}/getServiceById": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "Get service information",
                "description": "Returns service data",
                "operationId": "getServiceById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Service id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Service"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/service/{id}/updateServiceById": {
            "put": {
                "tags": [
                    "Services"
                ],
                "summary": "Update existing service",
                "description": "Returns updated service data",
                "operationId": "updateService",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Service id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateServiceRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Service"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/service/{id}/deleteServiceById": {
            "delete": {
                "tags": [
                    "Services"
                ],
                "summary": "Delete existing service",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteService",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Service id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Service": {
                "title": "Service",
                "description": "Service model",
                "properties": {
                    "title": {
                        "title": "title",
                        "description": "title of the Service model",
                        "example": ""
                    },
                    "description": {
                        "title": "description",
                        "description": "description of the Service model",
                        "example": ""
                    },
                    "icon": {
                        "title": "icon",
                        "description": "icon of the Service model",
                        "example": ""
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Service"
                }
            },
...
            "ServiceResource": {
                "title": "ServiceResource",
                "description": "Service resource",
                "properties": {
                    "data": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Service"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ServiceResource"
                }
            },
...
            "StoreServiceRequest": {
                "title": "Store Service request",
                "description": "Store Service request body data",
                "required": [
                    "title",
                    "description",
                    "icon"
                ],
                "properties": {
                    "title": {
                        "title": "title",
                        "description": "title of the new Service",
                        "type": "string",
                        "example": "Your example value here"
                    },
                    "description": {
                        "title": "description",
                        "description": "description of the new Service",
                        "type": "string",
                        "example": "Your example value here"
                    },
                    "icon": {
                        "title": "icon",
                        "description": "icon of the new Service",
                        "type": "string",
                        "example": "Your example value here"
                    }
                },
                "type": "object"
            },
...
            "UpdateServiceRequest": {
                "title": "UpdateService request",
                "description": "UpdateService request body data",
                "required": [
                    "title",
                    "description",
                    "icon"
                ],
                "properties": {
                    "title": {
                        "title": "title",
                        "description": "title of the new Service",
                        "type": "string",
                        "example": "Your example value here"
                    },
                    "description": {
                        "title": "description",
                        "description": "description of the new Service",
                        "type": "string",
                        "example": "Your example value here"
                    },
                    "icon": {
                        "title": "icon",
                        "description": "icon of the new Service",
                        "type": "string",
                        "example": "Your example value here"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}